Major refactoring and reformatting of settings.py
